Move Events test helpers to test package
Previously the helpers in pkg/reconciler/events/event_test.go were
exported publicly as part of the pkg/reconciler/events package.

This commit moves the helpers to the dedicated test package. The
helpers' signatures have been updated to prevent a circular dependency with
the pkg/reconciler/events/cloudevent package.
