+++ pkg/reconciler/events/event_test.go
@@ -24,6 +24,7 @@ import (
 	"github.com/tektoncd/pipeline/pkg/apis/config"
 	"github.com/tektoncd/pipeline/pkg/apis/pipeline/v1beta1"
 	"github.com/tektoncd/pipeline/pkg/reconciler/events/cloudevent"
+	eventstest "github.com/tektoncd/pipeline/test/events"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/tools/record"
@@ -142,7 +143,7 @@ func TestSendKubernetesEvents(t *testing.T) {
 		tr := &corev1.Pod{}
 		sendKubernetesEvents(fr, ts.before, ts.after, tr)
 
-		err := cloudevent.CheckEvents(t, fr, ts.name, ts.wantEvents)
+		err := eventstest.CheckEventsOrdered(t, fr.Events, ts.name, ts.wantEvents)
 		if err != nil {
 			t.Errorf(err.Error())
 		}
@@ -169,7 +170,7 @@ func TestEmitError(t *testing.T) {
 		tr := &corev1.Pod{}
 		EmitError(fr, ts.err, tr)
 
-		err := cloudevent.CheckEvents(t, fr, ts.name, ts.wantEvents)
+		err := eventstest.CheckEventsOrdered(t, fr.Events, ts.name, ts.wantEvents)
 		if err != nil {
 			t.Errorf(err.Error())
 		}
@@ -234,10 +235,10 @@ func TestEmit(t *testing.T) {
 
 		recorder := controller.GetEventRecorder(ctx).(*record.FakeRecorder)
 		Emit(ctx, nil, after, object)
-		if err := cloudevent.CheckEvents(t, recorder, tc.name, tc.wantEvents); err != nil {
+		if err := eventstest.CheckEventsOrdered(t, recorder.Events, tc.name, tc.wantEvents); err != nil {
 			t.Fatalf(err.Error())
 		}
-		if err := cloudevent.CheckCloudEvents(t, &fakeClient, tc.name, tc.wantCloudEvents); err != nil {
+		if err := eventstest.CheckEventsUnordered(t, fakeClient.Events, tc.name, tc.wantCloudEvents); err != nil {
 			t.Fatalf(err.Error())
 		}
 	}
